# software_SurtiTelas.Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SurtiTelas API",
    "version": "1.0.0",
    "description": "Backend API para la plataforma SurtiTelas - Manufactura y comercialización de prendas de vestir",
    "contact": {
      "name": "SurtiTelas Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/reports/sales": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Sales report",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sales aggregated report"
          }
        }
      }
    },
    "/reports/inventory": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Inventory report",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory aggregated report"
          }
        }
      }
    },
    "/reports/production": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Production report",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Production aggregated report"
          }
        }
      }
    },
    "/reports/users": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Users report",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users aggregated report"
          }
        }
      }
    },
...
  },
...
}